Relabelled mydecomposer to decomposer, decomposer to decomposer_base; removed instantiation of decomposer_base in subclass!
